feat: remove utils and add brownbaglast command
