+++ lib/google/youtube.js
@@ -21,15 +21,15 @@ const getStreams = async ({ max = 5, status = 'upcoming', pageToken = '' } = {})
   return streamsData.items || []
 }
 
-const getLastStream = async ({ max = 5, status = 'upcoming', pageToken = '' } = {}) => {
+const getLastStream = async ({ max = 1, status = 'upcoming', pageToken = '' } = {}) => {
   const streams = await getStreamsData({ max, status, pageToken })
 
   if (streams.nextPageToken) {
-    await getLastStream({ max, status, pageToken: streams?.nextPageToken })
-    return
+    const nextPageStreams = await getLastStream({ max, status, pageToken: streams?.nextPageToken })
+    return nextPageStreams || []
+  } else {
+    return streams.items || []
   }
-
-  return streams.items?.at(-1) || null
 }
 
 module.exports = { getStreams, getLastStream }
