+++ lib/utils.js
@@ -1,5 +1,3 @@
-const { DateTime } = require('luxon')
-
 const URL_RE = /https?:\/\/(www\.)?[-a-zA-Z0-9@:%._+~#=]{1,256}\.[a-zA-Z0-9()]{1,6}\b([-a-zA-Z0-9()@:%_+.~#?&//=]*)/g
 
 const isAllowedStreamYardUser = (userId) => {
@@ -14,21 +12,4 @@ const getDayName = (dateTime, locale = 'pt-br') => {
   return date.toLocaleDateString(locale, { weekday: 'long' }).toUpperCase()
 }
 
-const brownBagMessageList = (stream = null) => {
-  if (!stream) return []
-
-  const date = DateTime.fromISO(stream.snippet.scheduledStartTime, { locale: 'pt-br', zone: 'America/Sao_Paulo' })
-  const dateStr = date.toLocaleString(DateTime.DATETIME_MED)
-
-  return [
-    `**${stream.snippet.title}**`,
-    `_${dateStr}_`,
-    `https://youtube.com/watch?v=${stream.id}`,
-    '',
-    stream.snippet.description
-      ? preventUrlExpansion(stream.snippet.description).replace(/\r/gm, '').replace(/^/gm, '> ')
-      : null
-  ]
-}
-
-module.exports = { isAllowedStreamYardUser, preventUrlExpansion, getDayName, brownBagMessageList }
+module.exports = { isAllowedStreamYardUser, preventUrlExpansion, getDayName }
